refactor: move haversine function to utils module and clean up inference script
