+++ src/inference.py
@@ -1,49 +1,12 @@
 import logging
 
 import hydra
-import numpy as np
-import pandas as pd
 from hydra.utils import instantiate
 from omegaconf import DictConfig
 from prettytable import PrettyTable
 
 from data.dataset import load_test_dataset
-from tools.utils import get_kakao_lat_lng
-
-
-def haversine(
-    reviewer_lat: float, reviewer_lon: float, diner_lat: pd.Series, diner_lon: pd.Series
-) -> np.ndarray:
-    """
-    Compute the great-circle distance between a single point (lat1, lon1) and multiple points (lat2, lon2)
-    using the Haversine formula in a vectorized way.
-
-    Args:
-        reviewer_lat (float): Latitude of the reviewer.
-        reviewer_lon (float): Longitude of the reviewer.
-        diner_lat (pd.Series): Latitude of the diners.
-        diner_lon (pd.Series): Longitude of the diners.
-
-    Returns:
-        np.ndarray: Array of distances.
-    """
-    # Convert degrees to radians
-    reviewer_lat, reviewer_lon = np.radians(reviewer_lat), np.radians(reviewer_lon)
-    diner_lat, diner_lon = np.radians(diner_lat), np.radians(diner_lon)
-
-    # Haversine formula
-    dlat = diner_lat - reviewer_lat
-    dlon = diner_lon - reviewer_lon
-    a = (
-        np.sin(dlat / 2) ** 2
-        + np.cos(reviewer_lat) * np.cos(diner_lat) * np.sin(dlon / 2) ** 2
-    )
-    c = 2 * np.arctan2(np.sqrt(a), np.sqrt(1 - a))
-
-    # Earth's radius in kilometers
-    radius = 6371.0
-
-    return radius * c
+from tools.utils import get_kakao_lat_lng, haversine
 
 
 @hydra.main(config_path="../config/", config_name="predict", version_base="1.3.1")
